# server/api/swagger.yaml
...
swagger: '2.0'
info:
  version: '0.0.3'
  title: 'eCom Checkout API'
host: 'adidas.com'
basePath: '/api/checkout'
schemes:
  - 'https'
consumes:
  - 'application/json'
produces:
  - 'application/json; charset=utf-8'
security:
  - Authorization: []
  - Session Cookie: []
paths:
...
  /baskets:
    post:
      tags:
        - 'Basket'
      description: 'creates a new basket'
      operationId: 'basketsPOST'
      parameters: []
      responses:
        '201':
          description: 'Success'
          schema:
            $ref: '#/definitions/Basket'
        '400':
          description: 'error'
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: 'server error, api is not able to manage error'
      x-swagger-router-controller: 'Basket'
  /baskets/{basket_id}:
    get:
      tags:
        - 'Basket'
      description: 'DEPRECATED - returns the cart identified by the id to the caller'
      operationId: 'getCart'
      parameters:
        - name: 'basket_id'
          in: 'path'
          description: 'It represents basket unique identifier'
          required: true
          type: 'string'
          pattern: '^[a-z0-9]{26}$'
      responses:
        '200':
          description: 'Success'
          schema:
            $ref: '#/definitions/Basket'
        '400':
          description: 'error'
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: 'server error, api is not able to manage error'
      x-swagger-router-controller: 'Basket'
    delete:
      tags:
        - 'Basket'
      description: 'delete the given basket.'
      operationId: 'basketsBasket_idDELETE'
      parameters:
        - name: 'basket_id'
          in: 'path'
          description: 'It represents basket unique identifier'
          required: true
          type: 'string'
          pattern: '^[a-z0-9]{26}$'
      responses:
        '200':
          description: 'The basket was deleted successfully.'
        '400':
          description:
            "Indicates that the customer assigned to the basket does not\
            \ match the verified customer represented by the JWT token."
        '500':
          description: 'server error, api is not able to manage error'
      x-swagger-router-controller: 'Basket'
    patch:
      tags:
        - 'Basket'
      description: 'modifies some basket properties with the given content'
      operationId: 'basketsBasket_idPATCH'
      parameters:
        - name: 'basket_id'
          in: 'path'
          description: 'It represents basket unique identifier'
          required: true
          type: 'string'
          pattern: '^[a-z0-9]{26}$'
        - in: 'body'
          name: 'basketModification'
          required: true
          schema:
            $ref: '#/definitions/BasketModification'
      responses:
        '200':
          description: 'Success'
          schema:
            $ref: '#/definitions/Basket'
        '400':
          description: 'error'
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: 'server error, api is not able to manage error'
      x-swagger-router-controller: 'Basket'
  /baskets/{basket_id}/items:
    post:
      tags:
        - 'Basket'
      description: 'Allows the user to add a product to the cart'
      operationId: 'addToCart'
      parameters:
        - name: 'Content-Type'
          in: 'header'
          required: true
          type: 'string'
          enum:
            - 'application/json'
        - name: 'basket_id'
          in: 'path'
          description: 'It represents basket unique identifier'
          required: true
          type: 'string'
          pattern: '^[a-z0-9]{26}$'
        - in: 'body'
          name: 'products'
          description: 'The product to be added'
          required: true
          schema:
            $ref: '#/definitions/MultipleProductToAdd'
      responses:
        '200':
          description: 'Success'
          schema:
            $ref: '#/definitions/Basket'
        '400':
          description: 'error'
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: 'server error, api is not able to manage error'
      x-swagger-router-controller: 'Basket'
  /baskets/{basket_id}/items/{item_id}:
    delete:
      tags:
        - 'Basket'
      description: 'Allows the user to remove product from the cart'
      operationId: 'basketsBasket_idItemsItem_idDELETE'
      parameters:
        - name: 'basket_id'
          in: 'path'
          description: 'It represents basket unique identifier'
          required: true
          type: 'string'
          pattern: '^[a-z0-9]{26}$'
        - name: 'item_id'
          in: 'path'
          description: 'identify unique product line item id in the cart'
          required: true
          type: 'string'
          x-example: 'item_id'
      responses:
        '200':
          description: 'Success'
          schema:
            $ref: '#/definitions/Basket'
        '400':
          description: 'error'
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: 'server error, api is not able to manage error'
      x-swagger-router-controller: 'Basket'
    patch:
      tags:
        - 'Basket'
      description: 'Allow the user to update product line item of given basket'
      operationId: 'basketsBasket_idItemsItem_idPATCH'
      parameters:
        - name: 'basket_id'
          in: 'path'
          description: 'It represents basket unique identifier'
          required: true
          type: 'string'
          pattern: '^[a-z0-9]{26}$'
        - name: 'item_id'
          in: 'path'
          description: 'identify unique product line item id in the cart'
          required: true
          type: 'string'
          x-example: 'item_id'
        - in: 'body'
          name: 'product'
          description: 'The product to be updated'
          required: true
          schema:
            $ref: '#/definitions/ProductToAdd'
      responses:
        '200':
          description: 'Success'
          schema:
            $ref: '#/definitions/Basket'
        '400':
          description: 'error'
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: 'server error, api is not able to manage error'
      x-swagger-router-controller: 'Basket'
securityDefinitions:
  Authorization:
    description:
      "Authorization header from add to basket response to be used on every\
      \ subsequent request. If not provided, the API will generate it."
    type: 'basic'
  Session Cookie:
    description:
      "SFCC cookies from add to basket response (dwsid* and dwsecuretoken*)\
      \ to be sent on every subsequent request. If not provided, the API will generate\
      \ them."
    type: 'apiKey'
    name: 'SFCC Session Cookie'
    in: 'header'
definitions:
  Customer:
    properties:
      customerId:
        type: 'string'
        description: "Customer's unique id of SFCC"
      name:
        type: 'string'
        description: "Customer's name"
      email:
        description: 'It represents customer email address'
        $ref: '#/definitions/CustomerEmail'
      encryptedEmail:
        type: 'string'
        description: 'Encrypted version of customer email. Used for analytics purposes'
      receiveSmsUpdates:
        type: 'boolean'
        description:
          "Should the customer receive SMS notifications. Only enabled\
          \ for CnC delivery."
      isLoggedIn:
        type: 'boolean'
        description: 'Defines if this customer is logged in or not.'
    example:
      receiveSmsUpdates: true
      customerId: 'customerId'
      name: 'name'
      encryptedEmail: 'encryptedEmail'
      isLoggedIn: true
      email: {}
...
  MultipleProductToAdd:
    type: 'array'
    items:
      $ref: '#/definitions/ProductToAdd'
  ProductToAdd:
    type: 'object'
    required:
      - 'productId'
    properties:
      productId:
        description: 'The unique SKU'
        $ref: '#/definitions/ProductId'
      quantity:
        type: 'number'
        description: 'The amount of product to be added to cart. By default is 1'
      recipe:
        description: 'The recipe specification for personalized products'
        $ref: '#/definitions/Recipe'
      recipeId:
        type: 'string'
        description:
          "The unique receipe id sent from miconfigurator. It is used to\
          \ validate personalization recipe"
      customizationRecipeId:
        type: 'string'
        description:
          "The unique receipe id sent from miconfigurator. It is used to\
          \ validate customization recipe"
      masterProductId:
        description: 'TheUnique master product id.'
        $ref: '#/definitions/MasterProductId'
      displaySize:
        type: 'string'
        description: 'The country specific size value.'
      captchaResponse:
        type: 'string'
        description: 'google captcha response'
      segmentationId:
        type: 'string'
        description:
          "Segmentation string used to define the customer group for the\
          \ basket."
    example:
      captchaResponse: 'captchaResponse'
      quantity: 0.80082819046101150206595775671303272247314453125
      productId: {}
      customizationRecipeId: 'customizationRecipeId'
      masterProductId: {}
      recipe:
        quantity: 6.02745618307040320615897144307382404804229736328125
        articleNumber: 'articleNumber'
        selectedOptions: {}
        technicalSize: 'technicalSize'
        countrySize: 'countrySize'
        sku: 'sku'
      displaySize: 'displaySize'
      segmentationId: 'segmentationId'
      recipeId: 'recipeId'
...
  Basket:
    type: 'object'
    required:
      - 'basketId'
      - 'currency'
      - 'modifiedDate'
      - 'pricing'
      - 'resourceState'
      - 'taxationPolicy'
      - 'totalProductCount'
    properties:
      creationDate:
        description: 'It represents  date-time of Basket creation'
        $ref: '#/definitions/DateTime'
      modifiedDate:
        description: 'It represent last update date-time of Basket'
        $ref: '#/definitions/DateTime'
      customer:
        type: 'object'
        description: 'Customer related information'
        allOf:
          - $ref: '#/definitions/Customer'
          - type: 'object'
            properties: {}
      totalProductCount:
        type: 'integer'
        description: 'It shows total product count in basket'
      basketId:
        description: 'The unique identifier for the basket.'
        $ref: '#/definitions/BasketId'
      resourceState:
        type: 'string'
        description:
          "The resource_state represents the server side state of specific\
          \ resource. To use the resource state for optimistic locking you have to\
          \ include the resource state from the last server response in your next\
          \ state changing (DELETE, PATCH, POST, PUT) request. You should always prefer\
          \ to use the _resource_state property in the body. In case the requested\
          \ API doesn't have a body use the x-dw-resource-state header. Whenever a\
          \ resource state is part of a client's request, OCAPI verifies it by comparing\
          \ the given value with the server resource state. If both resource states\
          \ are equal the operation is executed, otherwise an HTTP 409 ResourceStateConflictException\
          \ fault is returned. In case of a create request (e.g. with PUT) where the\
          \ resource is not expected to be existing (with not_exists state), the returned\
          \ fault is an HTTP 409 ResourceAlreadyExistsException instead."
      taxationPolicy:
        type: 'string'
        description:
          "The taxation policy used for the basket. It will determine if\
          \ the prices in the basket are net or gross"
        enum:
          - 'net'
          - 'gross'
      currency:
        type: 'string'
        description: 'Currency information of price like EUR'
        pattern: '^[A-Z]{3}$'
      shipmentList:
        type: 'array'
        description: "It's a collection of shipment."
        items:
          $ref: '#/definitions/Shipment'
      billingAddress:
        description:
          "The billing address. This property is part of basket checkout\
          \ information only."
        $ref: '#/definitions/BillingAddress'
      pickupPoint:
        description: 'Pickup point information if a pickup point is set.'
        $ref: '#/definitions/PickupPoint'
      shippingAddress:
        description:
          "The collection of shipping address. Basket can have one shipping\
          \ addresses where products can be dispatched to."
        $ref: '#/definitions/ShippingAddress'
      shippingAddressValidation:
        description:
          "Address validation result based on the response from validation\
          \ service"
        $ref: '#/definitions/AddressValidationResult'
      pricing:
        description: 'The line item container pricing information'
        $ref: '#/definitions/LineItemContainerPricing'
      discountList:
        type: 'array'
        description: 'The collection of discount. Basket can have multiple discount.'
        items:
          $ref: '#/definitions/Discount'
      couponList:
        type: 'array'
        description:
          "The collection of coupon. It contains of any product, shipping\
          \ or order releated coupon."
        items:
          $ref: '#/definitions/Coupon'
      paymentInstrumentList:
        type: 'array'
        description: 'The collection of payment instruments'
        items:
          $ref: '#/definitions/PaymentInstrument'
      taxBreakDown:
        $ref: '#/definitions/Basket_taxBreakDown'
      taxBreakdownList:
        type: 'array'
        description: "The collection of Tax Breakdown field. It's used in CA market"
        items:
          $ref: '#/definitions/TaxBreakdown'
      messageList:
        type: 'array'
        items:
          $ref: '#/definitions/Message'
      checkoutId:
        type: 'string'
        description:
          "The checkout id associated with this basket for payment-related\
          \ operations"
      flashTimeLeft:
        type: 'number'
        description:
          "The time left, in milliseconds, since the most recently added\
          \ flash product was reserved"
      taxCalculationMissing:
        type: 'boolean'
        description:
          "true when tax calculation is missing, traditionally happens\
          \ when the address is wrong."
      segmentationId:
        type: 'string'
        description: 'string used to deliver information about ongoing AB tests.'
    example:
      basketId: '00000000000000000000000000'
      pickupPoint:
        cncStore:
          country: 'GR'
          telephoneNumber: '+358 10 123 1123'
          distance: 1.113520056782399958450469057424925267696380615234375
          city: 'Athens'
          postalCode: '1000AB'
          latitude: 37.97681879999999665642462787218391895294189453125
          daysToWaitForDelivery: 3
          deliveryWindow:
            from: 1
            to: 1
          street: 'Ermou Street 50'
          name: 'adidas Store Athens, Ermou Street,'
          openingHours:
            sunday:
              endHours: 18
              startHours: 9
              startMinutes: 0
              endMinutes: 30
            saturday:
              endHours: 18
              startHours: 9
              startMinutes: 0
              endMinutes: 30
            tuesday:
              endHours: 18
              startHours: 9
              startMinutes: 0
              endMinutes: 30
            wednesday:
              endHours: 18
              startHours: 9
              startMinutes: 0
              endMinutes: 30
            thursday:
              endHours: 18
              startHours: 9
              startMinutes: 0
              endMinutes: 30
            friday:
              endHours: 18
              startHours: 9
              startMinutes: 0
              endMinutes: 30
            monday:
              endHours: 18
              startHours: 9
              startMinutes: 0
              endMinutes: 30
          id: 'GR342248,'
          state: 'MOW'
          longitude: 23.72885409999999950514393276534974575042724609375
        pudoStore:
          country: 'GR'
          telephoneNumber: '+358 10 123 1123'
          distance: 1.113520056782399958450469057424925267696380615234375
          city: 'Athens'
          postalCode: '1000AB'
          latitude: 37.97681879999999665642462787218391895294189453125
          daysToWaitForDelivery: 3
          deliveryWindow:
            from: 1
            to: 1
          street: 'Ermou Street 50'
          name: 'adidas Store Athens, Ermou Street,'
          openingHours:
            sunday:
              endHours: 18
              startHours: 9
              startMinutes: 0
              endMinutes: 30
            saturday:
              endHours: 18
              startHours: 9
              startMinutes: 0
              endMinutes: 30
            tuesday:
              endHours: 18
              startHours: 9
              startMinutes: 0
              endMinutes: 30
            wednesday:
              endHours: 18
              startHours: 9
              startMinutes: 0
              endMinutes: 30
            thursday:
              endHours: 18
              startHours: 9
              startMinutes: 0
              endMinutes: 30
            friday:
              endHours: 18
              startHours: 9
              startMinutes: 0
              endMinutes: 30
            monday:
              endHours: 18
              startHours: 9
              startMinutes: 0
              endMinutes: 30
          id: 'GR342248,'
          state: 'MOW'
          longitude: 23.72885409999999950514393276534974575042724609375
      discountList:
        - calloutMessage: 'calloutMessage'
          couponItemId: 'couponItemId'
          name: 'name'
          id: 'id'
        - calloutMessage: 'calloutMessage'
          couponItemId: 'couponItemId'
          name: 'name'
          id: 'id'
      taxBreakDown:
        PST: 2.027123023002321833274663731572218239307403564453125
        QST: 7.3862819483858839220147274318151175975799560546875
        HST: 4.1456080298839363962315474054776132106781005859375
        GST: 3.61607674925191080461672754609026014804840087890625
      creationDate: {}
      paymentInstrumentList:
        - expirationYear: 7.061401241503109105224211816675961017608642578125
          lastFour: 'lastFour'
          maskedGiftCardNumber: 'maskedGiftCardNumber'
          balance: 9.301444243932575517419536481611430644989013671875
          paymentMethodId: 'paymentMethodId'
          cardType: 'cardType'
          holder: 'holder'
          id: 'id'
          expirationMonth: 2.3021358869347654518833223846741020679473876953125
        - expirationYear: 7.061401241503109105224211816675961017608642578125
          lastFour: 'lastFour'
          maskedGiftCardNumber: 'maskedGiftCardNumber'
          balance: 9.301444243932575517419536481611430644989013671875
          paymentMethodId: 'paymentMethodId'
          cardType: 'cardType'
          holder: 'holder'
          id: 'id'
          expirationMonth: 2.3021358869347654518833223846741020679473876953125
      taxCalculationMissing: true
      totalProductCount: 0
      shipmentList:
        - shippingOnDate: {}
          shippingLineItem: ''
          shipmentId: 'shipmentId'
          productLineItemList:
            - discountList:
                - calloutMessage: 'calloutMessage'
                  couponItemId: 'couponItemId'
                  name: 'name'
                  id: 'id'
                - calloutMessage: 'calloutMessage'
                  couponItemId: 'couponItemId'
                  name: 'name'
                  id: 'id'
              quantity: 60
              productId: {}
              gender: 'gender'
              color: 'color'
              canonicalProductName: 'canonicalProductName'
              availableStock: 5
              allowedActions:
                edit: true
                delete: true
                moveToWishlist: true
              personalizationFields: '{}'
              productName: 'productName'
              maxQuantityAllowed: 1
              itemId: 'itemId'
              productImage: 'productImage'
              size: 'size'
              editLinkCustomizableProduct: 'editLinkCustomizableProduct'
              isBonusProduct: true
              lastAdded: true
              isFlashProduct: true
              pricing:
                baseUnitPrice: {}
              productType: 'INLINE'
            - discountList:
                - calloutMessage: 'calloutMessage'
                  couponItemId: 'couponItemId'
                  name: 'name'
                  id: 'id'
                - calloutMessage: 'calloutMessage'
                  couponItemId: 'couponItemId'
                  name: 'name'
                  id: 'id'
              quantity: 60
              productId: {}
              gender: 'gender'
              color: 'color'
              canonicalProductName: 'canonicalProductName'
              availableStock: 5
              allowedActions:
                edit: true
                delete: true
                moveToWishlist: true
              personalizationFields: '{}'
              productName: 'productName'
              maxQuantityAllowed: 1
              itemId: 'itemId'
              productImage: 'productImage'
              size: 'size'
              editLinkCustomizableProduct: 'editLinkCustomizableProduct'
              isBonusProduct: true
              lastAdded: true
              isFlashProduct: true
              pricing:
                baseUnitPrice: {}
              productType: 'INLINE'
          shipmentType: 'inline'
        - shippingOnDate: {}
          shippingLineItem: ''
          shipmentId: 'shipmentId'
          productLineItemList:
            - discountList:
                - calloutMessage: 'calloutMessage'
                  couponItemId: 'couponItemId'
                  name: 'name'
                  id: 'id'
                - calloutMessage: 'calloutMessage'
                  couponItemId: 'couponItemId'
                  name: 'name'
                  id: 'id'
              quantity: 60
              productId: {}
              gender: 'gender'
              color: 'color'
              canonicalProductName: 'canonicalProductName'
              availableStock: 5
              allowedActions:
                edit: true
                delete: true
                moveToWishlist: true
              personalizationFields: '{}'
              productName: 'productName'
              maxQuantityAllowed: 1
              itemId: 'itemId'
              productImage: 'productImage'
              size: 'size'
              editLinkCustomizableProduct: 'editLinkCustomizableProduct'
              isBonusProduct: true
              lastAdded: true
              isFlashProduct: true
              pricing:
                baseUnitPrice: {}
              productType: 'INLINE'
            - discountList:
                - calloutMessage: 'calloutMessage'
                  couponItemId: 'couponItemId'
                  name: 'name'
                  id: 'id'
                - calloutMessage: 'calloutMessage'
                  couponItemId: 'couponItemId'
                  name: 'name'
                  id: 'id'
              quantity: 60
              productId: {}
              gender: 'gender'
              color: 'color'
              canonicalProductName: 'canonicalProductName'
              availableStock: 5
              allowedActions:
                edit: true
                delete: true
                moveToWishlist: true
              personalizationFields: '{}'
              productName: 'productName'
              maxQuantityAllowed: 1
              itemId: 'itemId'
              productImage: 'productImage'
              size: 'size'
              editLinkCustomizableProduct: 'editLinkCustomizableProduct'
              isBonusProduct: true
              lastAdded: true
              isFlashProduct: true
              pricing:
                baseUnitPrice: {}
              productType: 'INLINE'
          shipmentType: 'inline'
      taxationPolicy: 'net'
      shippingAddressValidation:
        jobToken: 'jobToken'
        isValid: true
        jobTokenRemainingUsage: 5.63737665663332876420099637471139430999755859375
        suggestionList:
          - zipCode: 'zipCode'
            province: 'province'
            formattedAddress: 'formattedAddress'
            city: 'city'
            street: 'street'
            houseNumber: 'houseNumber'
          - zipCode: 'zipCode'
            province: 'province'
            formattedAddress: 'formattedAddress'
            city: 'city'
            street: 'street'
            houseNumber: 'houseNumber'
        isCorrected: true
      messageList:
        - messageType: 'INFO'
          details:
            availableQuantity: 1.489415909854170383397331534069962799549102783203125
            articleNumber: 'articleNumber'
            shipmentId: 'shipmentId'
            requestedQuantity: 1.024645700144157789424070870154537260532379150390625
            outOfStockLetters:
              - 'outOfStockLetters'
              - 'outOfStockLetters'
            outOfStockNumbers:
              - 'outOfStockNumbers'
              - 'outOfStockNumbers'
            badgeMissing: true
            maximumQuantity: 1.231513536777255612975068288506008684635162353515625
          type: 'type'
        - messageType: 'INFO'
          details:
            availableQuantity: 1.489415909854170383397331534069962799549102783203125
            articleNumber: 'articleNumber'
            shipmentId: 'shipmentId'
            requestedQuantity: 1.024645700144157789424070870154537260532379150390625
            outOfStockLetters:
              - 'outOfStockLetters'
              - 'outOfStockLetters'
            outOfStockNumbers:
              - 'outOfStockNumbers'
              - 'outOfStockNumbers'
            badgeMissing: true
            maximumQuantity: 1.231513536777255612975068288506008684635162353515625
          type: 'type'
      couponList:
        - valid: true
          id: 'id'
          message: 'message'
          couponCode: 'couponCode'
          statusCode: 'coupon_code_already_in_basket'
        - valid: true
          id: 'id'
          message: 'message'
          couponCode: 'couponCode'
          statusCode: 'coupon_code_already_in_basket'
      resourceState: 'resourceState'
      shippingAddress:
        country: 'country'
        lastName: 'lastName'
        address2: 'address2'
        city: 'city'
        address1: 'address1'
        island: 'island'
        companyName: 'companyName'
        houseNumber: 'houseNumber'
        zipcode: 'zipcode'
        firstName: 'firstName'
        countyProvince: 'countyProvince'
        phoneNumber: 'phoneNumber'
        suburb: 'suburb'
        id: 'id'
        useAsBillingAddress: true
        apartmentNumber: 'apartmentNumber'
      flashTimeLeft: 6.8468526983526398765889098285697400569915771484375
      currency: 'currency'
      billingAddress:
        country: 'country'
        lastName: 'lastName'
        address2: 'address2'
        city: 'city'
        address1: 'address1'
        island: 'island'
        companyName: 'companyName'
        houseNumber: 'houseNumber'
        zipcode: 'zipcode'
        firstName: 'firstName'
        countyProvince: 'countyProvince'
        suburb: 'suburb'
        stateCode: {}
        id: 'id'
        apartmentNumber: 'apartmentNumber'
      checkoutId: 'checkoutId'
      segmentationId: 'segmentationId'
      pricing: {}
      taxBreakdownList:
        - name: 'name'
          value: 'value'
        - name: 'name'
          value: 'value'
      customer: ''
  BasketModification:
    type: 'object'
    properties:
      billingAddress:
        description: 'The billing address'
        $ref: '#/definitions/BillingAddress'
      shippingAddress:
        description: 'The shipping address'
        $ref: '#/definitions/ShippingAddress'
      customer:
        description: 'Customer related information'
        $ref: '#/definitions/Customer'
      methodList:
        type: 'array'
        description: 'The shipping methods to be set for this basket'
        items:
          type: 'object'
          allOf:
            - $ref: '#/definitions/ShippingMethod'
            - type: 'object'
              properties: {}
      newsletterSubscription:
        type: 'boolean'
        description: 'A flag to enable or not newsletter subscription for this user'
      pickupPoint:
        type: 'object'
        description: 'Set pickup point for e.g. CnC'
        allOf:
          - $ref: '#/definitions/CnCStoreModification'
          - $ref: '#/definitions/PUDOModification'
      segmentationId:
        type: 'string'
        description: 'string used to deliver information about ongoing AB tests.'
    example:
      pickupPoint: ''
      newsletterSubscription: true
      shippingAddress:
        country: 'country'
        lastName: 'lastName'
        address2: 'address2'
        city: 'city'
        address1: 'address1'
        island: 'island'
        companyName: 'companyName'
        houseNumber: 'houseNumber'
        zipcode: 'zipcode'
        firstName: 'firstName'
        countyProvince: 'countyProvince'
        phoneNumber: 'phoneNumber'
        suburb: 'suburb'
        id: 'id'
        useAsBillingAddress: true
        apartmentNumber: 'apartmentNumber'
      billingAddress:
        country: 'country'
        lastName: 'lastName'
        address2: 'address2'
        city: 'city'
        address1: 'address1'
        island: 'island'
        companyName: 'companyName'
        houseNumber: 'houseNumber'
        zipcode: 'zipcode'
        firstName: 'firstName'
        countyProvince: 'countyProvince'
        suburb: 'suburb'
        stateCode: {}
        id: 'id'
        apartmentNumber: 'apartmentNumber'
      methodList:
        - ''
        - ''
      segmentationId: 'segmentationId'
      customer:
        receiveSmsUpdates: true
        customerId: 'customerId'
        name: 'name'
        encryptedEmail: 'encryptedEmail'
        isLoggedIn: true
        email: {}
...
  BillingAddress:
    properties:
      id:
        type: 'string'
        description: 'The unique address id'
      address1:
        type: 'string'
        description: 'The address line 1 information'
      address2:
        type: 'string'
        description: 'The address line 2 information'
      city:
        type: 'string'
        description: 'City information'
      country:
        type: 'string'
        description: 'Country information'
      zipcode:
        type: 'string'
        description: 'Zipcode information'
      stateCode:
        description: 'The two digit code for a state'
        $ref: '#/definitions/StateCode'
      firstName:
        type: 'string'
        description: 'First name information'
      lastName:
        type: 'string'
        description: 'Last name information'
      countyProvince:
        type: 'string'
        description: 'County province information'
      companyName:
        type: 'string'
        description: 'Company name information'
      apartmentNumber:
        type: 'string'
        description: 'Apartment number information'
      houseNumber:
        type: 'string'
        description: 'House number information'
      suburb:
        type: 'string'
        description: 'Suburb information'
      island:
        type: 'string'
        description: 'Island information'
    description: 'Required Fields are market specific and defined via configurations'
    example:
      country: 'country'
      lastName: 'lastName'
      address2: 'address2'
      city: 'city'
      address1: 'address1'
      island: 'island'
      companyName: 'companyName'
      houseNumber: 'houseNumber'
      zipcode: 'zipcode'
      firstName: 'firstName'
      countyProvince: 'countyProvince'
      suburb: 'suburb'
      stateCode: {}
      id: 'id'
      apartmentNumber: 'apartmentNumber'
...
  BasketId:
    type: 'string'
    pattern: '^[a-z0-9]{26}$'
    x-example: '00000000000000000000000000'
...
  ErrorResponse:
    type: 'object'
    required:
      - 'message'
    properties:
      message:
        type: 'string'
        description: 'A human-readable message about the error'
      errorCode:
        type: 'string'
        description:
          "A code message in camelCase describing a error category or type\
          \ (ie. BasketNotFound)"
      messageList:
        description: 'Additional information provided by SFCC OCAPI response'
      details:
        $ref: '#/definitions/ErrorResponse_details'
...
  DateTime:
    type: 'string'
    description: 'ISO8601 Timestamp'
...